Refactor all operations
Tests must be the finishing thing..
