+++ crates/fhir-sdk/src/client/fhir/mod.rs
@@ -1,6 +1,7 @@
 //! FHIR API access methods (version specific).
 
 mod crud;
+mod operations;
 mod paging;
 mod patch;
 mod search_params;
